Include serverless in dev dependencies
Serverless is already installed as a peer dependency of several plugins.

Unfortunately this does not cause dependabot to keep the package version up to date, leading to unexpected behaviour when deploying with serverless.

Including the framework as a dev dependency should ensure we always have an up to date version installed for plugins to use.